Move `shared_examples` and `matcher` into specs.
